:broom: Tidying code before overhaul for geography setup
